Link to ndc search page from autocomplete search
