+++ app/javascript/app/data/initial-state.js
@@ -4,7 +4,7 @@ import { initialState as countrySDGLinkages } from 'components/country-ndc-sdg-l
 import { initialState as countryNDC } from 'pages/ndc-country';
 import { initialState as countryNDCFull } from 'pages/ndc-country-full';
 import { initialState as NDCCompare } from 'pages/ndc-compare';
-import { initialState as search } from 'pages/search';
+import { initialState as ndcSearch } from 'pages/ndc-search';
 import { initialState as countrySelect } from 'components/countries-select';
 import { initialState as autocompleteSearch } from 'components/autocomplete-search';
 import { initialState as countries } from 'providers/countries-provider';
@@ -19,5 +19,5 @@ export default {
   NDCCompare,
   countrySelect,
   autocompleteSearch,
-  search
+  ndcSearch
 };
