+++ app/javascript/app/pages/ndc-search/ndc-search-reducers.js
