+++ app/javascript/app/pages/ndc-search/ndc-search-selectors.js
@@ -2,7 +2,10 @@ import { createSelector } from 'reselect';
 
 const getResultsData = state => state.data;
 
-export const getSearchResults = createSelector(getResultsData, results => results);
+export const getSearchResults = createSelector(
+  getResultsData,
+  results => results
+);
 
 export default {
   getSearchResults
