+++ app/javascript/app/pages/ndc-search/ndc-search-styles.scss
