+++ app/javascript/app/reducers.js
@@ -9,14 +9,19 @@ import { reducers as countrySDGLinkagesReducers } from 'components/country-ndc-s
 import { reducers as countryNDCReducers } from 'pages/ndc-country';
 import { reducers as countryNDCFullReducers } from 'pages/ndc-country-full';
 import { reducers as NDCCompareReducers } from 'pages/ndc-compare';
-import { reducers as searchReducers } from 'pages/search';
+import { reducers as ndcSearchReducers } from 'pages/ndc-search';
 import { reducers as mapReducers } from 'components/map';
 import initialState from './data/initial-state';
 import allActions from './actions';
 
 export default combineReducers({
   ndcs: handleActions('ndcs', allActions, NDCSReducers, initialState),
-  search: handleActions('search', allActions, searchReducers, initialState),
+  ndcSearch: handleActions(
+    'search',
+    allActions,
+    ndcSearchReducers,
+    initialState
+  ),
   countries: handleActions(
     'countries',
     allActions,
