+++ app/javascript/app/routes.js
@@ -11,7 +11,7 @@ import NDCCountryFull from 'pages/ndc-country-full';
 import NDCCompare from 'pages/ndc-compare';
 import CountryIndex from 'pages/country-index';
 import Country from 'pages/country';
-import Search from 'pages/search';
+import NDCSearch from 'pages/ndc-search';
 import GHGEmissions from 'components/country-ghg-emissions';
 import NDCSDGLinkages from 'components/country-ndc-sdg-linkages';
 import error from 'pages/error';
@@ -88,9 +88,9 @@ export default [
         component: CountryIndex
       },
       {
-        path: '/search',
+        path: '/ndc-search',
         exact: true,
-        component: Search
+        component: NDCSearch
       },
       {
         path: '/error-page',
